Updates 1-22-2017
Added two additional configuration options to autonomous: whether to
collect the third ball or not, and how far forward to travel to collect
the ball. Note: going forward to collect the third ball (as opposed to
collecting it without moving forwards by careful robot positioning) is
not currently being used due to driving accuracy issues, so
thirdBallDistance should never be changed from zero.

Wrote and fine-tuned program to collect third ball from in front of our
partner's robot and continue on with the rest of the autonomous, for
Blue only.

Added moving forward before aligning with wall prior to sensing for the
near beacon, to avoid getting stuck or hitting the corner vortex (for
Blue only)