Updates 1-23-2017 3
Removed some TeleOp code that was no longer useful.

Fixed/Improved shooter control in TeleOp. New controls: Pressing A (the
button) once turns on the shooter. Pressing A again, if the shooter has
been running for 1.2 seconds by this point, opens the door. Otherwise,
it waits for the 1.2 seconds to elapse, and then opens the door. Once
the door has been opened, pressing A again turns off the shooter and
closes the door.